+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 0.3.3 (10.03.2016)
+* add `inesita watch`
+* add `inesita build dir parameters`
+
 ## 0.3.2 (05.03.2016)
 * add `app_dist` and `app_dev` directory support
-* add `redirect_to` to route options
+* add `redirect_to` to `route` options
 
 ## 0.3.1 (05.01.2016)
 * add params hash to `router.url_for`